implement random.invidious.io
automatically redirects requests to a random, healthy instance. the user
may specify which region the instance should be from by setting the
iv-region query parameter and the minimum 30 day uptime in percent with
iv-health query parameter. if no instance matching these parameters can
be found, a 404 error is returned instead.

note that this behaviour is only enabled for the magic hostname
'random.invidious.io' in order to not send traffic to instances unless
explicitly requested.